Cria arquivos dos componentes Navbar, Layout, Footer, Table, CardBox
