+++ src/components/CardBox.jsx/CardBox.jsx
@@ -1,0 +1,6 @@
+export default function CardBox() {
+    return (
+        <>
+        </>
+    )
+}
