--- src/components/Footer.jsx
@@ -1,7 +1,0 @@
-export default function Footer() {
-    return (
-        <>
-            <h3>Capacita +</h3>
-        </>
-    )
-}
