+++ src/components/Layout/Layout.jsx
@@ -1,0 +1,16 @@
+import './style.css'
+import Navbar from '../Navbar/Navbar';
+import Footer from '../Footer/Footer';
+import { Outlet } from 'react-router-dom';
+
+export default function Layout() {
+  return (
+    <div className='app-container'>
+      <Navbar />
+      <main className='main-content'>
+        <Outlet />
+      </main>
+      <Footer />
+    </div>
+  );
+}
