+++ src/components/Layout/style.css
@@ -1,0 +1,13 @@
+.app-container {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  min-width: 320px;
+  overflow-x: hidden;
+}
+
+.main-content {
+  flex: 1;
+  padding: 16px;
+  box-sizing: border-box;
+}
