+++ src/components/Table/Table.jsx
@@ -1,0 +1,6 @@
+export default function Table() {
+    return (
+        <>
+        </>
+    )
+}
